[QA] we must plan huge range of value for score ;)

+++ Assets/Scripts/Behaviour/GameStateSO.cs
@@ -9,6 +9,6 @@ namespace Behaviour
     {
         public float currentWheelSpeed = 0.5f;
         public float currentFame = 0.5f;
-        public float score = 0.0f;
+        public double score = 0.0f;
     }
 }
